+++ helpers.js
@@ -3,10 +3,6 @@ const childProcess = require("child_process");
 const { access } = require("fs/promises");
 const fs = require("fs");
 
-const {
-  PYTEST_CLI_NAME,
-} = require("./consts.json");
-
 const exec = util.promisify(childProcess.exec);
 
 async function assertPathExistence(path) {
@@ -17,28 +13,7 @@ async function assertPathExistence(path) {
   }
 }
 
-function chooseCommand(command, jsonReport) {
-  if (command.substring(0, PYTEST_CLI_NAME.length) !== PYTEST_CLI_NAME) {
-    throw new Error(`Command must begin with "${PYTEST_CLI_NAME}".`);
-  }
-  if (command.includes("--json-report") || jsonReport === undefined) {
-    return command;
-  }
-
-  switch (jsonReport) {
-    case "none":
-      return command;
-    case "full":
-      return `${PYTEST_CLI_NAME} --json-report${command.substring(PYTEST_CLI_NAME.length)}`;
-    case "summary":
-      return `${PYTEST_CLI_NAME} --json-report --json-report-summary${command.substring(PYTEST_CLI_NAME.length)}`;
-    default:
-      throw new Error("Parameter JSON Report is not properly specified as \"none\", \"full\", or \"summary\"");
-  }
-}
-
 module.exports = {
   assertPathExistence,
   exec,
-  chooseCommand,
 };
